[COLAB-2918]-[comment]: split up controller into three classes according to the interfaces; swap RequestMapping with GetMapping; remove instance() method on client interfaces; add StaticInjector for clients within the comment-client module; remove all             <artifactId>spring-cloud-starter-feign</artifactId> entries in poms
