+++ microservices/clients/comment-client/src/main/java/org/xcolab/client/comment/pojo/ICategory.java
@@ -3,8 +3,7 @@ package org.xcolab.client.comment.pojo;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import org.codehaus.jackson.map.annotate.JsonDeserialize;
 
-import org.xcolab.client.comment.CategoryClient;
-import org.xcolab.client.comment.ThreadClient;
+import org.xcolab.client.comment.StaticInjectorComment;
 import org.xcolab.client.comment.exceptions.CategoryGroupNotFoundException;
 import org.xcolab.client.comment.exceptions.KeyReferenceException;
 import org.xcolab.client.comment.pojo.tables.pojos.Category;
@@ -54,7 +53,7 @@ public interface ICategory {
 
     @JsonIgnore
     default List<IThread> getThreads(ThreadSortColumn sortColumn, boolean ascending) {
-        return ThreadClient.instance().listThreads(0, Integer.MAX_VALUE,
+        return StaticInjectorComment.getThreadClient().listThreads(0, Integer.MAX_VALUE,
                 getId(), null, sortColumn, ascending);
     }
 
@@ -63,7 +62,7 @@ public interface ICategory {
         final Long groupId = getGroupId();
         if (groupId != null && groupId > 0) {
             try {
-                return CategoryClient.instance().getCategoryGroup(groupId);
+                return StaticInjectorComment.getCategoryClient().getCategoryGroup(groupId);
             } catch (CategoryGroupNotFoundException e) {
                 throw new KeyReferenceException(e);
             }
