+++ microservices/clients/comment-client/src/main/java/org/xcolab/client/comment/pojo/ICategoryGroup.java
@@ -3,8 +3,7 @@ package org.xcolab.client.comment.pojo;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 
-import org.xcolab.client.comment.CategoryClient;
-import org.xcolab.client.comment.ThreadClient;
+import org.xcolab.client.comment.StaticInjectorComment;
 import org.xcolab.client.comment.pojo.tables.pojos.CategoryGroup;
 import org.xcolab.client.comment.util.ThreadSortColumn;
 
@@ -31,13 +30,14 @@ public interface ICategoryGroup {
 
     @JsonIgnore
     default List<IThread> getThreads(ThreadSortColumn sortColumn, boolean ascending) {
-        return ThreadClient.instance().listThreads(0, Integer.MAX_VALUE,
+        return StaticInjectorComment.getThreadClient().listThreads(0, Integer.MAX_VALUE,
                 null, getId(), sortColumn, ascending);
     }
 
     @JsonIgnore
     default List<ICategory> getCategories() {
-        return CategoryClient.instance().listCategories(0, Integer.MAX_VALUE, getId());
+        return StaticInjectorComment.getCategoryClient()
+                .listCategories(0, Integer.MAX_VALUE, getId());
     }
 
     @JsonIgnore
