+++ microservices/clients/content-client/pom.xml
@@ -15,12 +15,6 @@
 
     <dependencies>
         <dependency>
-            <groupId>org.springframework.cloud</groupId>
-            <artifactId>spring-cloud-starter-feign</artifactId>
-            <version>${spring-cloud.feign}</version>
-        </dependency>
-
-        <dependency>
             <groupId>org.xcolab</groupId>
             <artifactId>xcolab-utils</artifactId>
             <version>${project.version}</version>
