+++ pom.xml
@@ -29,7 +29,6 @@
         <!-- Spring versions -->
         <spring-boot.version>2.0.7.RELEASE</spring-boot.version>
         <spring-cloud.release-train>Finchley.SR2</spring-cloud.release-train>
-        <spring-cloud.feign>1.4.6.RELEASE</spring-cloud.feign>
 
         <!-- Database -->
         <jooq.version>3.9.2</jooq.version>
